requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mutagen==1.47.0
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
PyYAML==6.0.1
requests==2.32.3
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.5.27